+++ CHANGELOG.md
@@ -101,6 +101,7 @@
 - Add routes to manage actions and conditions on a content rule @robgietema
 - Add summary method to content rules @robgietema
 - Add content rule assignments endpoints @robgietema
+- Add assignments to content rule endpoints @robgietema
 
 ### Bugfix